requirements_ai.txt
# AI Model Training and Testing Requirements
# For Crop Disease Detection (SDG 2 - Zero Hunger)

# Core ML/AI Libraries
tensorflow>=2.10.0,<3.0.0
tensorflow-gpu>=2.10.0,<3.0.0  # Optional: for GPU acceleration
numpy>=1.21.0
scikit-learn>=1.1.0
opencv-python>=4.6.0
Pillow>=9.0.0

# Model Export and Optimization
tf2onnx>=1.13.0
onnx>=1.12.0
onnxruntime>=1.12.0

# Data Processing and Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0

# Testing and Validation
pytest>=7.0.0
pytest-cov>=3.0.0
psutil>=5.8.0  # For memory monitoring

# Dataset Management
kaggle>=1.5.12  # For downloading PlantVillage dataset
requests>=2.28.0

# Utilities
tqdm>=4.64.0  # Progress bars
pyyaml>=6.0
python-dotenv>=0.19.0

# Optional: Jupyter for interactive development
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Model monitoring and logging
tensorboard>=2.10.0
mlflow>=1.28.0

# Optional: Advanced optimization
tensorflow-model-optimization>=0.7.0  # For quantization
tensorflow-addons>=0.19.0  # For additional layers

# Development and Testing
black>=22.0.0  # Code formatting
flake8>=5.0.0  # Linting
mypy>=0.991  # Type checking

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Performance Monitoring
memory-profiler>=0.60.0
line-profiler>=3.5.0

# Mobile Deployment Support
tensorflowjs>=3.18.0  # For web deployment 